backend/backend/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0

# Audio processing
openai-whisper>=20231117  # Requires FFmpeg: Run python install_ffmpeg.py
pandas>=2.0.0
numpy>=1.24.0

# Text analysis
spacy>=3.5.0
textstat>=0.7.3

# HTTP client
requests>=2.28.0
httpx>=0.24.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0

# Utilities
tqdm>=4.65.0

# Ensure spaCy English model is installed
# Run: python -m spacy download en_core_web_sm
